+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 - `SuperwallOptions.localResources` now accepts UIImage's from xcasset files, e.g. `UIImage(named: "my-image")`.
 - Exposes abandoned transaction product params in audience filters.
 
+### Fixes
+
+- Sanitizes email user attribute.
+
 ## 4.15.0
 
 ### Enhancements